+++ test/test-c-editconfig.sh
@@ -21,9 +21,6 @@ test -d $CFD || mkdir -p $CFD
 
 fyang=$dir/myyang.yang
 
-# source IMG/USER etc
-. ./site.sh
-
 # Common NACM scripts
 . ./nacm.sh