+++ test/test-c-simulated.sh
@@ -22,9 +22,6 @@ test -d $CFD || mkdir -p $CFD
 
 fyang=$dir/myyang.yang
 
-# source IMG/USER etc
-. ./site.sh
-
 # Common NACM scripts
 . ./nacm.sh